+++ examples/nuxt/plugins/walletManager.client.ts
@@ -26,7 +26,8 @@ export default defineNuxtPlugin((nuxtApp) => {
       {
         id: WalletId.MAGIC,
         options: { apiKey: 'pk_live_D17FD8D89621B5F3' }
-      }
+      },
+      WalletId.MNEMONIC
     ],
     defaultNetwork: NetworkId.TESTNET
   })