+++ examples/react-ts/src/App.tsx
@@ -28,7 +28,8 @@ const walletManager = new WalletManager({
     {
       id: WalletId.MAGIC,
       options: { apiKey: 'pk_live_D17FD8D89621B5F3' }
-    }
+    },
+    WalletId.MNEMONIC
   ],
   defaultNetwork: NetworkId.TESTNET
 })