+++ packages/use-wallet-vue/vite.config.ts
@@ -1,15 +1,7 @@
-import { defineConfig } from 'vitest/config'
+import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
 import vueJsx from '@vitejs/plugin-vue-jsx'
 
 export default defineConfig({
-  plugins: [vue(), vueJsx()],
-  test: {
-    name: 'use-wallet-vue',
-    dir: './src',
-    watch: false,
-    environment: 'jsdom',
-    setupFiles: ['./vitest.setup.ts'],
-    globals: true
-  }
+  plugins: [vue(), vueJsx()]
 })