+++ packages/use-wallet/src/index.ts
@@ -1,7 +1,7 @@
 export { LogLevel } from './logger'
 export { WalletManager, WalletManagerConfig, WalletManagerOptions } from './manager'
 export { NetworkId, DEFAULT_NETWORKS } from './network'
-export { State, WalletState, DEFAULT_STATE } from './store'
+export { State, WalletState, ManagerStatus, DEFAULT_STATE } from './store'
 export { StorageAdapter } from './storage'
 export { webpackFallback } from './webpack'
 export * from './wallets'