+++ packages/use-wallet/vitest.config.ts
@@ -2,6 +2,9 @@ import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   test: {
+    name: 'use-wallet',
+    dir: './src',
+    watch: false,
     globals: true,
     setupFiles: './setupTests.ts'
   }